+++ src/schema/FroidSchema.ts
@@ -20,6 +20,7 @@ import {ObjectTypeNode} from './types';
 import {
   CONTRACT_DIRECTIVE_NAME,
   DEFAULT_FEDERATION_LINK_IMPORTS,
+  Directive,
   DirectiveName,
   FED2_OPT_IN_URL,
   FED2_VERSION_PREFIX,
@@ -184,10 +185,15 @@ export class FroidSchema {
     this.generateFroidDependencies();
 
     // build schema
+    const imports = [...DEFAULT_FEDERATION_LINK_IMPORTS];
+    if (this.hasShareableTypes()) {
+      imports.push(Directive.Shareable);
+    }
+
     this.froidAst = sortDocumentAst({
       kind: Kind.DOCUMENT,
       definitions: [
-        this.createLinkSchemaExtension(),
+        this.createLinkSchemaExtension(imports),
         this.createQueryDefinition(),
         this.createNodeInterface(),
         ...this.createCustomReturnTypes(),
@@ -242,7 +248,14 @@ export class FroidSchema {
         )
       );
 
-      if (isException || !FroidSchema.isEntity(node)) {
+      const isShareableValueType = node.directives?.some(
+        (directive) => directive.name.value === DirectiveName.Shareable
+      );
+
+      if (
+        isException ||
+        (!FroidSchema.isEntity(node) && !isShareableValueType)
+      ) {
         return;
       }
 
@@ -861,5 +874,16 @@ export class FroidSchema {
    */
   private static parseSchema(schema: string): DocumentNode {
     return parse(schema, {noLocation: true});
+  }
+
+  /**
+   * Checks if there are any shareable types in the FROID object types.
+   *
+   * @returns {boolean} True if there are shareable types, false otherwise.
+   */
+  private hasShareableTypes(): boolean {
+    return Object.values(this.froidObjectTypes).some(
+      (objectType) => objectType.isShareable
+    );
   }
 }