+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.2.3] - 2025-09-29
+
+### Fix
+
+- The `FroidSchema` class does not include all shareable value type fields found
+  across subgraphs.
+
 ## [v3.2.2] - 2024-08-20
 
 ### Fixed